Fix review query bugs causing startup/runtime errors
- ReviewRepository: sumRatingByProductId was an invalid Spring Data derived
  query; replaced with @Query using COALESCE(SUM(...))
- OrderRepository: COUNT > 0 boolean SELECT not valid JPQL; changed to
  COUNT return with > 0 check in service layer

+++ product-service/src/main/java/com/ecommerce/product/service/ReviewService.java
@@ -50,7 +50,7 @@ public class ReviewService {
     public ReviewSummaryResponse getSummary(Long productId, Long userId) {
         int total = reviewRepository.countByProductId(productId);
         double avg = total == 0 ? 0.0
-                : reviewRepository.sumRatingByProductId(productId) / total;
+                : (double) reviewRepository.sumRatingByProductId(productId) / total;
 
         List<ReviewResponse> last5 = reviewRepository
                 .findTop5ByProductIdOrderByCreatedAtDesc(productId)
